==> 5_signals/0_non_responders/1_figures/2_volcanoes.ipynb <==
source(paste0(dirname(dirname(dirname(getwd()))),'/map.r'))
source(paste0(HELP_DIR, "shortcuts.r"))
source(paste0(HELP_DIR, "helpers.r"))

library(forcats)
#library(wesanderson)

### 0 - Read exhaustive analysis

fe <- fread(paste0(SHARE_DIR, "1_run_fishers_exact.csv"))

remove_last <- function(i) i[-length(i)]

options(repr.plot.height= 6, repr.plot.width = 8)

base <- 
fe %>% 
 rw() %>% mu( raw = paste(remove_last(strsplit(feature, "_")[[1]]), collapse = "_")) %>% ug() %>% 
 gb(cohortGo, raw, fisher_pval) %>% mu( direction = factor(direction, levels = c("Response Associated", "Non-Response Associated"))) %>% ug() 

#head(base)

labels <- base %>% fi(or < 1) %>% gb(cohortGo) %>% mu(rk = row_number(fisher_pval)) %>%fi(rk == 1)

options(repr.plot.height= 6, repr.plot.width = 20)

base %>% 
 fi(grepl("Pan-Cancer", cohortGo)) %>% 
 ggplot( aes(y = -log10(fisher_pval), x = log(or))) + 
 geom_point() + 
 #scale_alpha_manual( values = alphas ) + 
 facet_wrap(~fct_reorder(cohortGo, fisher_pval, .fun = min), ncol = 5) + 
 labs( y = "-Log 10 (Fisher Exact P-value)", x = "Log Odds Response") + 
 go_theme + 
 ylim(0,11) + xlim(-5,5) + 
 geom_hline(yintercept = -log10(.05), type = "dashed") + 
 theme(legend.position = "none") + 
 geom_text(data = labels %>% fi(grepl("Pan-Cancer", cohortGo)), aes(label = feature), vjust = -1, color = "black", size = 3)
